feat(backend): create and list MultiSigAccount
This commit includes some important changes

* An initial structure for service and controllers. Now we have a way to pass the components down to service and controllers without the need to have a global instance of DB accessed directly via packages.
* Change package names to follow Go package naming conventions. See more: https://go.dev/blog/package-names
* A paginateutil that makes it easy to get the pagination params (limit and offset) from HTTP query and use it on services

### About Services

Service is a component that we can use to interact with an Entity. Ideally, it works with a single domain.

The service can receive dependencies like database, cache, or anything else. It's a good practice because we have testable components and an easy way to mock injected dependencies.

Example service packages:

* multisigaccountsvc
* transactionsvc (coming soon)
* signedtransactionsvc  (coming soon)

### About Controllers

Controllers is also a component, but it is used to handle HTTP requests and call the services.

Example controllers packages:

* multisigaccountctrl
* transactionctrl(coming soon)
* signedtransactionctrl  (coming soon)