refactor(backend): move zap logger config to viper config
Currently, we have a global logger variable that is used across all the backend.

Since loggers are dependencies (just like references to other components), the best practice is to instantiate the logger on the main and pass the reference to all the service that needs the logger.

Read more: https://peter.bourgon.org/go-best-practices-2016/#top-tip-10